requirements.txt
plotly
streamlit
matplotlib
streamlit_keplergl
numerize
pillow
